+++ src/main/java/Domus/DatasetUtils/ReadJsonMain.java
@@ -1,9 +1,8 @@
 package Domus.DatasetUtils;
 
-import Domus.DatasetUtils.DataserClass.Dataset;
+import Domus.DatasetUtils.DatasetClass.Dataset;
 import com.google.gson.Gson;
 
-import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.IOException;
 import java.io.Reader;
