+++ src/main/java/Domus/DomusTestDriverFullList.java
@@ -1,6 +1,6 @@
 package Domus;
 
-import Domus.DatasetUtils.DataserClass.Dataset;
+import Domus.DatasetUtils.DatasetClass.Dataset;
 import Domus.DatasetUtils.DomusRecord;
 
 public class DomusTestDriverFullList extends DomusTestDriver{
