+++ src/main/java/Domus/Experiments/TestDriverWMethodLstar.java
@@ -1,29 +1,20 @@
 package Domus.Experiments;
 
-import Domus.DatasetUtils.CustomGson;
-import Domus.DatasetUtils.DataserClass.Dataset;
+import Domus.DatasetUtils.DatasetClass.Dataset;
 import Domus.DatasetUtils.DomusRecord;
 import Domus.DomusOracle;
 import Domus.DomusTestDriver;
-import Domus.DomusWord;
 import Domus.VisualizeGraph;
-import com.google.gson.Gson;
 import de.learnlib.algorithms.lstar.dfa.ClassicLStarDFA;
 import de.learnlib.algorithms.lstar.dfa.ClassicLStarDFABuilder;
-import de.learnlib.algorithms.rivestschapire.RivestSchapireDFA;
 import de.learnlib.api.oracle.EquivalenceOracle;
 import de.learnlib.api.oracle.MembershipOracle;
 import de.learnlib.datastructure.observationtable.OTUtils;
-import de.learnlib.oracle.equivalence.DFAWMethodEQOracle;
 import de.learnlib.oracle.equivalence.DFAWpMethodEQOracle;
-import de.learnlib.oracle.equivalence.SampleSetEQOracle;
 import de.learnlib.util.Experiment;
-import de.learnlib.util.statistics.SimpleProfiler;
 import net.automatalib.automata.fsa.DFA;
 import net.automatalib.automata.fsa.impl.compact.CompactDFA;
-import net.automatalib.serialization.dot.GraphDOT;
 
-import java.io.FileWriter;
 import java.io.IOException;
 
 import static Domus.Experiments.ExperimentUtils.readJson;
